Add files via upload
1.Airlines_data 
Perform clustering (hierarchical, KMeans clustering, and DBSCAN) for the airline's data to obtain an optimum number of clusters. Draw the inferences from the clusters obtained.

The file EastWestAirlines contains information on passengers who belong to an airline’s frequent flier program. For each passenger, the data include information on their mileage history and on different ways they accrued or spent miles in the last year. The goal is to try to identify clusters of passengers that have similar characteristics for the purpose of targeting different segments for different types of mileage offers

2.Crime_data
Perform Clustering(Hierarchical, Kmeans & DBSCAN) for the crime data and identify the number of clusters formed and draw inferences.

Data Description: Murder -- Murder rates in different places of United States
Assault- Assault rate in different places of United States
Urban Pop - urban population in different places of United States
Rape - Rape rate in different places of United States